+++ crates/red_knot_python_semantic/src/stdlib.rs
